requirements.txt
# requirements.txt
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
great_expectations==0.18.0
mlflow==2.3.1
prefect==3.4.1
pyyaml==6.0
griffe==1.7.3
